Test isolation
Change the backend interface so setup() and teardown() are used in the
expected way; setup() should be run before each tests, and teardown()
should be run after each test. Backends implement setup() to restore the
storage to the initial condition to ensure test isolation.

Creating storage is done one using "create" and deleting it using
"delete".