Wait for udev events after creating a filesystem
In imageio tests on CentOS 8, we see this failure:

    [userstorage] INFO    Creating filesystem /var/tmp/imageio-storage/file-512-xfs-mount
    mount: /var/tmp/imageio-storage/file-512-xfs-mount: wrong fs type, bad
    option, bad superblock on /dev/loop2, missing codepage or helper
    program, or other error.

RHEL 8 docs[1] recommends to run:

    udevadm settle

after creating a filesystem, to wait for the system to register the new
device node.

The call should be harmless on older systems even if it is not required,
but now we require systemd-udev providing udevadm.

[1] https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/8/html/managing_file_systems/assembly_getting-started-with-xfs-managing-file-systems