Raise CreateError for expected errors
Mounting loop device fails randomly on some setups (CentOS 8 mock) when
running on host with older kernel (3.10). Previously we failed hard,
aborting main. Now we raise CreateError, so if the storage is optional,
we can continue to create the next backend. If we are lucky, most
storage will be created without failing the build.

I did not see mkfs failures yet, but it makes sense to treat such
failures in the same way.