backend: check exists before accessing
Add exists() call to check if the setup() has
been done correctly before yielding the current
storage while entering the context.

This adds pytest dependency to the backend, but most
projects already depend on pytest. If at some
point it is requested to remove this dependency,
it can be made optional.

Add a test to ensure that calling setup() without
create() will end with xfail.

Signed-off-by: Albert Esteve <aesteve@redhat.com>